Fix several -Wsign-compare compiler warnings
-Wextra showed the following [-Wsign-compare] warnings

src/parser.c: In function ‘ff_request_is_raw_http’:
src/parser.c:50:23: warning: comparison of integer expressions of different signedness: ‘int’ and ‘long unsigned int’ [-Wsign-compare]
   50 |     for (int i = 0; i < (sizeof(HTTP_METHODS) / sizeof(HTTP_METHODS[0])); i++)
      |                       ^

src/http.c: In function ‘ff_http_send_request_unencrypted’:
src/http.c:149:19: warning: comparison of integer expressions of different signedness: ‘int’ and ‘uint64_t’ {aka ‘long unsigned int’} [-Wsign-compare]
  149 |     } while (sent < request->payload_length);
      |                   ^
src/http.c:174:23: warning: comparison of integer expressions of different signedness: ‘int’ and ‘long unsigned int’ [-Wsign-compare]
  174 |     } while (received < sizeof(response) - 1);
      |                       ^

In all the above the simple solution was to just add casts to (int) as
necessary.

For example, with the last two, I'm not sure how likely (or indeed
feasible) it would be to send/receive more than 2GiB of data (INT_MAX)
in a single request.

+++ src/http.c
@@ -146,7 +146,7 @@ bool ff_http_send_request_unencrypted(struct ff_request *request, char *host_nam
         }
 
         sent += chunk;
-    } while (sent < request->payload_length);
+    } while (sent < (int)request->payload_length);
 
     ff_log(FF_DEBUG, "Finished sending request to %s over HTTP (%d bytes sent)", host_name, sent);
 
@@ -171,7 +171,7 @@ bool ff_http_send_request_unencrypted(struct ff_request *request, char *host_nam
         {
             break;
         }
-    } while (received < sizeof(response) - 1);
+    } while (received < (int)sizeof(response) - 1);
 
     ff_log(FF_DEBUG, "Finished receiving response from %s (%d bytes received)", host_name, received);
     size_t response_header_length = strchr((char *)response, '\n') - response;
